Change data type of HEPE to float
This is to accomodate values of accuracy
that exceed 65535

CRs-fixed: 449184

+++ loc_api/libloc_api_50001/loc_ulp.h
@@ -264,7 +264,7 @@ typedef struct  {
     /** longitude in degrees */
     double longitude;
     /** Horzizontal error estimate in meters */
-    uint16_t HEPE;
+    float HEPE;
 } UlpNetworkPosition;
 
 /** Represents access point information */
